+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.5.3 - 4-12-2020
+
+* Added new usage examples
+* Update documentation
+
+## 0.5.2 - 4-12-2020
+
+* Added selectedItemIconSize field in BottomBarTheme class
+* Remove ci/cd logic from travis to GithubActions
+
 ## 0.5.1 - 5-11-2020
 
 ### Created Widget tests